add gitlab package unifying related packages
