--- internal/data/md.mock.go
@@ -1,41 +1,0 @@
-package data
-
-const Content = `
-# Today’s Menu
-
-## Appetizers
-
-| Name        | Price | Notes                           |
-| ---         | ---   | ---                             |
-| Tsukemono   | $2    | Just an appetizer               |
-| Tomato Soup | $4    | Made with San Marzano tomatoes  |
-| Okonomiyaki | $4    | Takes a few minutes to make     |
-| Curry       | $3    | We can add squash if you’d like |
-
-## Seasonal Dishes
-
-| Name                 | Price | Notes              |
-| ---                  | ---   | ---                |
-| Steamed bitter melon | $2    | Not so bitter      |
-| Takoyaki             | $3    | Fun to eat         |
-| Winter squash        | $3    | Today it's pumpkin |
-
-## Desserts
-
-| Name         | Price | Notes                 |
-| ---          | ---   | ---                   |
-| Dorayaki     | $4    | Looks good on rabbits |
-| Banana Split | $5    | A classic             |
-| Cream Puff   | $3    | Pretty creamy!        |
-
-All our dishes are made in-house by Karen, our chef. Most of our ingredients
-are from our garden or the fish market down the street.
-
-Some famous people that have eaten here lately:
-
-* [x] René Redzepi
-* [x] David Chang
-* [ ] Jiro Ono (maybe some day)
-
-Bon appétit!
-`
