--- internal/data/mergerequest.mock.go
@@ -1,226 +1,0 @@
-package data
-
-import (
-	"time"
-
-	"github.com/felipeospina21/mrglab/internal/gql"
-)
-
-var GQLMergeRequestMock = gql.MergeRequestConnection{
-	Count: 1,
-	Edges: []gql.MergeRequestEdge{
-		{
-			Node: gql.MergeRequestNode{
-				IID:                 "1",
-				DiffHeadSha:         "some-sha",
-				Title:               "Mocked Title 1",
-				CreatedAt:           time.Now(),
-				UpdatedAt:           time.Now(),
-				Draft:               false,
-				Author:              gql.Author{Name: "Mock User"},
-				DetailedMergeStatus: "ci_still_running",
-				Conflicts:           true,
-				UserNotesCount:      0,
-				WebURL:              "",
-				Description:         "Dummy description",
-				ApprovalsRequired:   3,
-				DiffStatsSummary: gql.DiffStatsSummary{
-					Additions: 10,
-					Deletions: 50,
-					Changes:   100,
-					FileCount: 5,
-				},
-			},
-		},
-		{
-			Node: gql.MergeRequestNode{
-				IID:                 "2",
-				Title:               "Mocked Title 2",
-				CreatedAt:           time.Now(),
-				UpdatedAt:           time.Now(),
-				Draft:               false,
-				Author:              gql.Author{Name: "Mock User"},
-				DetailedMergeStatus: "blocked_status",
-				Conflicts:           true,
-				UserNotesCount:      10,
-				WebURL:              "",
-				Description:         "Dummy description",
-				ApprovalsRequired:   3,
-				DiffStatsSummary: gql.DiffStatsSummary{
-					Additions: 10,
-					Deletions: 50,
-					Changes:   100,
-					FileCount: 5,
-				},
-			},
-		},
-		{
-			Node: gql.MergeRequestNode{
-				IID:                 "3",
-				Title:               "Mocked Title 3",
-				CreatedAt:           time.Now(),
-				UpdatedAt:           time.Now(),
-				Draft:               false,
-				Author:              gql.Author{Name: "Mock User"},
-				DetailedMergeStatus: "conflict",
-				Conflicts:           true,
-				UserNotesCount:      0,
-				WebURL:              "",
-				Description:         "Dummy description",
-				ApprovalsRequired:   3,
-				DiffStatsSummary: gql.DiffStatsSummary{
-					Additions: 10,
-					Deletions: 50,
-					Changes:   100,
-					FileCount: 5,
-				},
-			},
-		},
-		{
-			Node: gql.MergeRequestNode{
-				IID:                 "4",
-				Title:               "Mocked Title 4",
-				CreatedAt:           time.Now(),
-				UpdatedAt:           time.Now(),
-				Draft:               false,
-				Author:              gql.Author{Name: "Mock User"},
-				DetailedMergeStatus: "mergeable",
-				Conflicts:           false,
-				UserNotesCount:      0,
-				WebURL:              "",
-				Description:         "Dummy description",
-				ApprovalsRequired:   1,
-				DiffStatsSummary: gql.DiffStatsSummary{
-					Additions: 10,
-					Deletions: 50,
-					Changes:   100,
-					FileCount: 5,
-				},
-				ApprovalState: gql.MergeRequestApprovalState{
-					Rules: []gql.ApprovalRule{
-						{
-							Name:              "Rule 1",
-							ApprovalsRequired: 1,
-							Approved:          false,
-							ApprovedBy: gql.ApprovedBy{
-								Nodes: []gql.ApprovedByNode{
-									{Name: "user"},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			Node: gql.MergeRequestNode{
-				IID:                 "5",
-				Title:               "Mocked Title 5",
-				CreatedAt:           time.Now(),
-				UpdatedAt:           time.Now(),
-				Draft:               false,
-				Author:              gql.Author{Name: "Mock User"},
-				DetailedMergeStatus: "external_status_checks",
-				Conflicts:           false,
-				UserNotesCount:      0,
-				WebURL:              "",
-				Description:         "Dummy description",
-				ApprovalsRequired:   3,
-				DiffStatsSummary: gql.DiffStatsSummary{
-					Additions: 10,
-					Deletions: 50,
-					Changes:   100,
-					FileCount: 5,
-				},
-			},
-		},
-		{
-			Node: gql.MergeRequestNode{
-				IID:                 "6",
-				Title:               "Mocked Title 6",
-				CreatedAt:           time.Now(),
-				UpdatedAt:           time.Now(),
-				Draft:               false,
-				Author:              gql.Author{Name: "Mock User"},
-				DetailedMergeStatus: "need_rebase",
-				Conflicts:           false,
-				UserNotesCount:      0,
-				WebURL:              "",
-				Description:         "Dummy description",
-				ApprovalsRequired:   3,
-				DiffStatsSummary: gql.DiffStatsSummary{
-					Additions: 10,
-					Deletions: 50,
-					Changes:   100,
-					FileCount: 5,
-				},
-			},
-		},
-		{
-			Node: gql.MergeRequestNode{
-				IID:                 "7",
-				Title:               "Mocked Title 7",
-				CreatedAt:           time.Now(),
-				UpdatedAt:           time.Now(),
-				Draft:               false,
-				Author:              gql.Author{Name: "Mock User"},
-				DetailedMergeStatus: "draft_status",
-				Conflicts:           false,
-				UserNotesCount:      0,
-				WebURL:              "",
-				Description:         "Dummy description",
-				ApprovalsRequired:   3,
-				DiffStatsSummary: gql.DiffStatsSummary{
-					Additions: 10,
-					Deletions: 50,
-					Changes:   100,
-					FileCount: 5,
-				},
-			},
-		},
-		{
-			Node: gql.MergeRequestNode{
-				IID:                 "8",
-				Title:               "Mocked Title 8",
-				CreatedAt:           time.Now(),
-				UpdatedAt:           time.Now(),
-				Draft:               false,
-				Author:              gql.Author{Name: "Mock User"},
-				DetailedMergeStatus: "checking",
-				Conflicts:           false,
-				UserNotesCount:      0,
-				WebURL:              "",
-				Description:         "Dummy description",
-				ApprovalsRequired:   3,
-				DiffStatsSummary: gql.DiffStatsSummary{
-					Additions: 10,
-					Deletions: 50,
-					Changes:   100,
-					FileCount: 5,
-				},
-			},
-		},
-		{
-			Node: gql.MergeRequestNode{
-				IID:                 "9",
-				Title:               "Mocked Title 9",
-				CreatedAt:           time.Now(),
-				UpdatedAt:           time.Now(),
-				Draft:               false,
-				Author:              gql.Author{Name: "Mock User"},
-				DetailedMergeStatus: "external_status_checks",
-				Conflicts:           false,
-				UserNotesCount:      0,
-				WebURL:              "",
-				Description:         "Dummy description",
-				ApprovalsRequired:   3,
-				DiffStatsSummary: gql.DiffStatsSummary{
-					Additions: 10,
-					Deletions: 50,
-					Changes:   100,
-					FileCount: 5,
-				},
-			},
-		},
-	},
-}
