--- internal/gql/entities.go
@@ -1,17 +1,0 @@
-package gql
-
-type Author struct {
-	Name string
-}
-
-type Labels struct {
-	Count int
-	Edges []struct {
-		Node struct {
-			Color     string
-			Title     string
-			TextColor string
-			ID        string
-		}
-	}
-}
