+++ internal/tui/components/projects/commands.go
@@ -2,14 +2,13 @@ package projects
 
 import (
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/felipeospina21/mrglab/internal/api"
-	"github.com/felipeospina21/mrglab/internal/gql"
+	"github.com/felipeospina21/mrglab/internal/gitlab"
 	"github.com/felipeospina21/mrglab/internal/tui"
 )
 
 func (m *Model) GetListCmd() tea.Cmd {
 	return func() tea.Msg {
-		mrs, err := api.GetProjectMergeRequestsGQL(m.ctx.SelectedProject.ID, gql.MergeRequestsQueryVariables{
+		mrs, err := m.client.GetProjectMergeRequests(m.ctx.SelectedProject.ID, gitlab.MergeRequestsQueryVariables{
 			State: "opened",
 		})
